refactor: CrewAttendances 초기화 시 파일 입력 메서드 컨트롤러로 분리
